refactor, fix api
refactor
fix clippy
fix docs
remove unused code
fix bytesfield index api flaw
