+++ src/indexer/merger.rs
@@ -130,14 +130,6 @@ impl TermOrdinalMapping {
     fn get_segment(&self, segment_ord: usize) -> &[TermOrdinal] {
         &(self.per_segment_new_term_ordinals[segment_ord])[..]
     }
-
-    fn max_term_ord(&self) -> TermOrdinal {
-        self.per_segment_new_term_ordinals
-            .iter()
-            .flat_map(|term_ordinals| term_ordinals.iter().max().cloned())
-            .max()
-            .unwrap_or_default()
-    }
 }
 
 struct DeltaComputer {
@@ -814,7 +806,7 @@ impl IndexMerger {
             doc_id_mapping,
             &reader_and_field_accessors,
         )?;
-        let mut serialize_vals = fast_field_serializer.new_bytes_fast_field_with_idx(field, 1);
+        let mut serialize_vals = fast_field_serializer.new_bytes_fast_field(field);
 
         for old_doc_addr in doc_id_mapping.iter_old_doc_addrs() {
             let bytes_reader = &reader_and_field_accessors[old_doc_addr.segment_ord as usize].1;
